Fix imports and move Song and SongCounter classes to model package
